Changed return value to NSUInteger to match NSObject
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/systempreferences/trunk@31624 72102866-910b-0410-8b05-ffd578937521

+++ Modules/TimeZone/MapView.m
@@ -1,6 +1,6 @@
 /* MapView.m
  *  
- * Copyright (C) 2005 Free Software Foundation, Inc.
+ * Copyright (C) 2005-2010 Free Software Foundation, Inc.
  *
  * Author: Enrico Sersale <enrico@imago.ro>
  * Date: December 2005
@@ -211,7 +211,7 @@ double myrintf(double a)
 	[super dealloc];
 }
 
-- (unsigned)hash
+- (NSUInteger)hash
 {
   return (zone != nil) ? [zone hash] : [super hash];
 }
